Separate CipherSuite/Dhkem classes dependent on external modules from the independent base classes.
